Revert merge which broke connections.
Revert "fix: use time properly for conversion between sec and ms"

This reverts commit b7292c6465899043b50c13c11c4d92e05d3952b1.

Revert "feat: add code to use revised lifetime"

This reverts commit 57bf70114859f7ea1e65843a83c89f41226ecdb5.

+++ NET-Core/LibUA/Types.cs
@@ -7059,7 +7059,7 @@ namespace LibUA
 
             public uint ChannelID { get; set; }
             public uint TokenID { get; set; }
-            public double TokenLifetime { get; set; }
+            public UInt32 TokenLifetime { get; set; }
             public DateTimeOffset TokenCreatedAt { get; set; }
 
             public uint? PrevChannelID { get; set; }
